Backend/mychatbot/requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
joblib==1.4.0
nltk==3.8.1
regex==2023.12.25
sqlparse==0.5.0
tqdm==4.66.2
tzdata==2024.1
